Rename from RBS::Diff::CLI to RBS::CLI::Diff
