+++ lib/rbs/cli.rb
@@ -9,6 +9,7 @@ require "stringio"
 module RBS
   class CLI
     autoload :ColoredIO, 'rbs/cli/colored_io'
+    autoload :Diff, 'rbs/cli/diff'
 
     class LibraryOptions
       attr_accessor :core_root
@@ -1391,7 +1392,7 @@ EOB
     end
 
     def run_diff(argv, library_options)
-      Diff::CLI.new(argv: argv, library_options: library_options, stdout: stdout, stderr: stderr).run
+      Diff.new(argv: argv, library_options: library_options, stdout: stdout, stderr: stderr).run
     end
   end
 end
