+++ sig/cli.rbs
@@ -28,18 +28,6 @@ module RBS
     attr_reader stdout: _IO
     attr_reader stderr: _IO
 
-    class ColoredIO
-      attr_reader stdout: CLI::_IO
-      def initialize: (stdout: CLI::_IO) -> void
-      def puts_green: (String) -> void
-      def puts_red: (String) -> void
-      def puts: (String) -> void
-
-      private def can_display_colors?: () -> bool
-      private def are_colors_supported?: () -> bool
-      private def are_colors_disabled?: () -> bool
-    end
-
     # The copy of `args` passed to `run`.
     #
     # `OptionParser#order!` mutates given `arg`, the `run_***` actions should read `original_args` to get the original commandline arguments.
