+++ sig/diff.rbs
@@ -1,27 +1,5 @@
 module RBS
   class Diff
-    class CLI
-      interface _Puts
-        def puts: (*untyped) -> void
-      end
-
-      @format: String
-      @stdout: _Puts
-      @stderr: _Puts
-      @diff: Diff
-
-      def initialize: (
-        argv: Array[String],
-        library_options: RBS::CLI::LibraryOptions,
-        ?stdout: _Puts,
-        ?stderr: _Puts,
-      ) -> void
-
-      def run: () -> void
-      def run_diff: () -> void
-      def run_markdown: () -> void
-    end
-
     @type_name: TypeName
     @library_options: RBS::CLI::LibraryOptions
     @after_path: Array[String]
